+++ CHANGELOG.md
@@ -19,10 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Handler for too long names
 - Exit button in lobby sends you back to main menu
 - Main menu button in results
+- Grayed out start game button for non-hosts
 
 ## Fixed
 - Host can't have an empty name
 - Game lock when 6 players are in the lobby
+- Can't start a new round when current one isn't finished yet
 
 ## 2024-07-12